# backend/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HTL Kahoot API",
        "description": "API documentation for the HTL Kahoot application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "List questions",
                "description": "Returns a paginated, filterable list of questions. Accessible by teachers, admins and superadmins.",
                "operationId": "475dfb2e0b5a1ff1de8bfa52421af77b",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "multiple_choice"
                        }
                    },
                    {
                        "name": "is_published",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "pool_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "type",
                                "is_published"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "with_trashed",
                        "in": "query",
                        "description": "Include soft-deleted questions (admin/superadmin only)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated question list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuestionList"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a question",
                "description": "Creates a new question and its first version atomically. Accessible by teachers, admins and superadmins.",
                "operationId": "4bc43023949407b69ca1f3dc27ab464d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "title"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "multiple_choice"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "What is the capital of France?"
                                    },
                                    "explanation": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "difficulty": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "nullable": true
                                    },
                                    "default_points": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 1000
                                    },
                                    "default_time_limit": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "nullable": true
                                    },
                                    "randomize_options": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "config": {
                                        "type": "object"
                                    },
                                    "answer_options": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "text": {
                                                    "type": "string",
                                                    "example": "Paris"
                                                },
                                                "is_correct": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "sort_order": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/questions/{id}/restore": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Restore a soft-deleted question",
                "description": "Restores a previously soft-deleted question. Admin and superadmin only.",
                "operationId": "5139444b107652865f7604706bdb009b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restored question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/questions/{id}/versions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "List question versions",
                "description": "Returns all versions of a question ordered by version number. Owner or admin/superadmin only.",
                "operationId": "b485c6e81fba2a93cec48ad2fb07e96e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of versions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/QuestionVersion"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/questions/{id}/publish": {
            "patch": {
                "tags": [
                    "Questions"
                ],
                "summary": "Toggle published status",
                "description": "Toggles the is_published flag of a question. Owner or admin/superadmin only.",
                "operationId": "7237e1ba5f441ac9573ae267d5d1338a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 87
                    }
                },
                "type": "object"
            },
...
            "QuestionList": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}